consensus_benchmark/consensus_benchmark/metrics.py
#!/usr/bin/env python3
"""
Metrics module for evaluating model responses.

This module provides functions to calculate various metrics for evaluating
the quality of LLM responses, including accuracy, factual consistency,
hallucination rates, and reasoning quality.
"""

import re
from typing import Dict, List, Any, Union, Optional
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity


def calculate_accuracy(responses: List[Dict], ground_truth: List[Dict]) -> float:
    """
    Calculate the accuracy of model responses compared to ground truth.
    
    Args:
        responses: List of response dictionaries
        ground_truth: List of ground truth dictionaries
        
    Returns:
        Accuracy score (0.0 to 1.0)
    """
    # In a real implementation, this would do semantic matching
    # For now, we'll return a simulated score
    return np.random.uniform(0.6, 0.9)


def calculate_factual_consistency(responses: List[Dict]) -> float:
    """
    Calculate the factual consistency of model responses.
    
    This evaluates if the response contains inconsistent or contradictory facts.
    
    Args:
        responses: List of response dictionaries
        
    Returns:
        Factual consistency score (0.0 to 1.0)
    """
    # In a real implementation, this would check for internal consistency
    # For now, we'll return a simulated score
    return np.random.uniform(0.7, 0.95)


def calculate_hallucination_rate(
    responses: List[Dict], 
    ground_truth: List[Dict]
) -> float:
    """
    Calculate the hallucination rate of model responses.
    
    Hallucinations are statements that are not grounded in provided context
    or are factually incorrect based on ground truth.
    
    Args:
        responses: List of response dictionaries
        ground_truth: List of ground truth dictionaries
        
    Returns:
        Hallucination rate (0.0 to 1.0, lower is better)
    """
    # In a real implementation, this would identify fabricated information
    # For now, we'll return a simulated score
    return np.random.uniform(0.05, 0.3)


def calculate_response_consistency(responses: List[Dict]) -> float:
    """
    Calculate the consistency of model responses.
    
    This evaluates if the model gives consistent answers to similar questions.
    
    Args:
        responses: List of response dictionaries
        
    Returns:
        Consistency score (0.0 to 1.0)
    """
    # In a real implementation, this would check consistency across similar prompts
    # For now, we'll return a simulated score
    return np.random.uniform(0.7, 0.9)


def calculate_reasoning_quality(responses: List[Dict]) -> float:
    """
    Calculate the reasoning quality of model responses.
    
    This evaluates the logical coherence and step-by-step reasoning.
    
    Args:
        responses: List of response dictionaries
        
    Returns:
        Reasoning quality score (0.0 to 1.0)
    """
    # In a real implementation, this would evaluate logical structure
    # For now, we'll return a simulated score
    return np.random.uniform(0.6, 0.9)


def detect_contradictions(text: str) -> List[Dict[str, str]]:
    """
    Detect contradictory statements within a text.
    
    Args:
        text: Input text to analyze
        
    Returns:
        List of detected contradictions

