consensus_benchmark/consensus_benchmark/requirements.txt
# Core dependencies
numpy>=1.22.0
pandas>=1.4.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.8.0

# LLM API interactions
openai>=1.0.0
anthropic>=0.5.0
langchain>=0.0.267

# Evaluation and metrics
rouge>=1.0.1
nltk>=3.7
scikit-learn>=1.0.0
bert-score>=0.3.12
sentence-transformers>=2.2.0

# Data processing
tqdm>=4.64.0
pyyaml>=6.0
jsonschema>=4.17.0

# Utilities
python-dotenv>=0.20.0
requests>=2.28.0

